#2996: stick with net.jpountz.lz4
