+++ factcast-grpc-lz4/pom.xml
@@ -51,8 +51,8 @@
       <type>test-jar</type>
     </dependency>
     <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-compress</artifactId>
+      <groupId>net.jpountz.lz4</groupId>
+      <artifactId>lz4</artifactId>
     </dependency>
     <dependency>
       <groupId>org.factcast</groupId>
