+++ factcast-grpc-lz4/src/main/java/org/factcast/grpc/lz4/Lz4GrpcServerCodec.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.server.grpc.codec;
+package org.factcast.grpc.lz4;
 
 import io.grpc.Codec;
 import java.io.InputStream;
