+++ factcast-grpc-lz4/src/test/java/org/factcast/grpc/lz4/Lz4GrpcClientCodecTest.java
@@ -13,18 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.factcast.server.grpc.codec;
+package org.factcast.grpc.lz4;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("deprecation")
-class Lz4GrpcServerCodecTest {
+class Lz4GrpcClientCodecTest {
 
-  final Lz4GrpcServerCodec uut = new Lz4GrpcServerCodec();
+  final Lz4GrpcClientCodec uut = new Lz4GrpcClientCodec();
 
   @Test
   void getMessageEncoding() {
