--- factcast-grpc-lz4/src/test/java/org/factcast/grpc/lz4/Lz4cCodecPerformanceTest.java
@@ -1,27 +1,0 @@
-/*
- * Copyright © 2017-2020 factcast.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.factcast.grpc.lz4;
-
-import io.grpc.Codec;
-import org.factcast.test.AbstractCompressionPerformanceTest;
-
-class Lz4cCodecPerformanceTest extends AbstractCompressionPerformanceTest {
-
-  @Override
-  protected Codec getCodecToTest() {
-    return new Lz4cGrpcClientCodec();
-  }
-}
