+++ factcast-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -6,12 +6,10 @@ org.factcast.spring.boot.autoconfigure.factus.FactusAutoConfiguration
 org.factcast.spring.boot.autoconfigure.factus.BinaryJacksonSnapshotSerializerAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.GrpcFactStoreAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.LZ4ClientAutoConfiguration
-org.factcast.spring.boot.autoconfigure.client.grpc.LZ4cClientAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.SnappyClientAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.SnappycClientAutoConfiguration
 org.factcast.spring.boot.autoconfigure.server.grpc.FactCastGrpcServerAutoConfiguration
 org.factcast.spring.boot.autoconfigure.server.grpc.LZ4ServerAutoConfiguration
-org.factcast.spring.boot.autoconfigure.server.grpc.LZ4cServerAutoConfiguration
 org.factcast.spring.boot.autoconfigure.server.grpc.SnappyServerAutoConfiguration
 org.factcast.spring.boot.autoconfigure.server.grpc.SnappycServerAutoConfiguration
 org.factcast.spring.boot.autoconfigure.server.FactCastServerCommonSecurityAutoConfiguration
